lab4/requirements.txt
fastapi
uvicorn
passlib[bcrypt]
sqlalchemy
psycopg2
pydantic
python-jose[cryptography]
python-multipart
pymongo
dnspython
